Use feh instead of nitrogen for backgrounds
This is a bit interesting; without completely validating this theory,
I was having issues on the external monitor where closing the last
window would not refresh the display (so it was hard to tell if the
window was gone or not).  Replacing with feh and I haven't had the
problem since.